requirements.txt
streamlit>=1.28.0
biopython>=1.81
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0 